--- jwp-basic-gradle/src/main/java/servlet/UpdateUserFormServlet.java
@@ -1,25 +1,0 @@
-package servlet;
-
-import db.DataBase;
-
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@WebServlet("/user/updateForm")
-public class UpdateUserFormServlet extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException{
-        resp.setCharacterEncoding("UTF-8");
-        resp.setContentType("text/html;charset=UTF-8");
-
-        req.setAttribute("user", DataBase.findUserById(req.getParameter("userId")));
-
-        RequestDispatcher dispatcher = req.getRequestDispatcher("/user/update.jsp");
-        dispatcher.forward(req, resp);
-    }
-}
