+++ jwp-basic-gradle/src/main/java/servlet/UpdateUserServlet.java
@@ -14,9 +14,20 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
-@WebServlet("/user/update")
+@WebServlet(value = { "/user/update", "/user/updateForm" })
 public class UpdateUserServlet extends HttpServlet {
     private static final Logger log = LoggerFactory.getLogger(UpdateUserServlet.class);
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException{
+        resp.setCharacterEncoding("UTF-8");
+        resp.setContentType("text/html;charset=UTF-8");
+
+        req.setAttribute("user", DataBase.findUserById(req.getParameter("userId")));
+
+        RequestDispatcher dispatcher = req.getRequestDispatcher("/user/update.jsp");
+        dispatcher.forward(req, resp);
+    }
 
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         resp.setCharacterEncoding("UTF-8");
